refactor: dataset delete button independent of DatasetList component
