+++ src/views/DatasetManager/components/DatasetList/DatasetList.js
@@ -1,13 +1,8 @@
 // Copyright (c) Cosmo Tech.
 // Licensed under the MIT license.
 import React, { useCallback, useEffect, useMemo, useRef, useState } from 'react';
-import { Trans, useTranslation } from 'react-i18next';
-import {
-  DeleteForever as DeleteForeverIcon,
-  Refresh as RefreshIcon,
-  Search as SearchIcon,
-  Error as ErrorIcon,
-} from '@mui/icons-material';
+import { useTranslation } from 'react-i18next';
+import { Refresh as RefreshIcon, Search as SearchIcon, Error as ErrorIcon } from '@mui/icons-material';
 import {
   Box,
   Card,
@@ -25,8 +20,8 @@ import { ResourceUtils } from '@cosmotech/core';
 import { DontAskAgainDialog, PermissionsGate, SearchBar } from '@cosmotech/ui';
 import { DATASET_SOURCE_TYPE, INGESTION_STATUS } from '../../../../services/config/ApiConstants';
 import { ACL_PERMISSIONS } from '../../../../services/config/accessControl';
-import { TwoActionsDialogService } from '../../../../services/twoActionsDialog/twoActionsDialogService';
 import { CreateDatasetButton } from '../CreateDatasetButton';
+import { DeleteDatasetButton } from '../DeleteDatasetButton/DeleteDatasetButton';
 import { ReuploadFileDatasetButton } from '../ReuploadFileDatasetButton';
 import { useDatasetList } from './DatasetListHook';
 
@@ -52,15 +47,8 @@ export const DatasetList = () => {
   };
 
   const [isRefreshConfirmationDialogOpen, setIsRefreshConfirmationDialogOpen] = useState(false);
-  const {
-    userPermissionsInCurrentOrganization,
-    datasets,
-    currentDataset,
-    selectDataset,
-    deleteDataset,
-    refreshDatasetById,
-    isDatasetCopyEnabledInWorkspace,
-  } = useDatasetList();
+  const { userPermissionsInCurrentOrganization, datasets, currentDataset, selectDataset, refreshDatasetById } =
+    useDatasetList();
 
   const sortedDatasetList = useMemo(() => {
     return ResourceUtils.getResourceTree(datasets);
@@ -92,46 +80,6 @@ export const DatasetList = () => {
   useEffect(() => {
     filterDatasets(searchString);
   }, [searchString, filterDatasets]);
-
-  const askConfirmationToDeleteDialog = useCallback(
-    async (event, dataset) => {
-      event.stopPropagation();
-      const impactedScenariosWarning = isDatasetCopyEnabledInWorkspace
-        ? ''
-        : ' ' + // Space character is here on purpose, to separate concatenated sentences in confirmation dialog body
-          t(
-            'commoncomponents.datasetmanager.dialogs.delete.impactedScenariosWarning',
-            'All the scenarios using this dataset will be impacted.'
-          );
-
-      const dialogProps = {
-        id: 'delete-dataset',
-        component: 'div',
-        labels: {
-          title: t('commoncomponents.datasetmanager.dialogs.delete.title', 'Delete dataset?'),
-          body: (
-            <Trans
-              i18nKey="commoncomponents.datasetmanager.dialogs.delete.body"
-              defaultValue="Do you really want to delete <i>{{datasetName}}</i>?
-                This action is irreversible.{{impactedScenariosWarning}}"
-              values={{ datasetName: dataset?.name, impactedScenariosWarning }}
-              shouldUnescape={true}
-            />
-          ),
-          button1: t('commoncomponents.datasetmanager.dialogs.cancel', 'Cancel'),
-          button2: t('commoncomponents.datasetmanager.dialogs.delete.confirmButton', 'Delete'),
-        },
-        button2Props: {
-          color: 'error',
-        },
-      };
-      const result = await TwoActionsDialogService.openDialog(dialogProps);
-      if (result === 2) {
-        deleteDataset(dataset?.id);
-      }
-    },
-    [t, deleteDataset, isDatasetCopyEnabledInWorkspace]
-  );
 
   const confirmAndRefreshDataset = useCallback((event, callbackFunction) => {
     event.stopPropagation();
@@ -208,18 +156,12 @@ export const DatasetList = () => {
             userPermissions={userPermissionsOnDataset}
             necessaryPermissions={[ACL_PERMISSIONS.DATASET.DELETE]}
           >
-            <IconButton
-              onClick={(event) => askConfirmationToDeleteDialog(event, dataset)}
-              data-cy={`dataset-delete-button-${dataset.id}`}
-              disabled={dataset.ingestionStatus === INGESTION_STATUS.PENDING}
-            >
-              <DeleteForeverIcon />
-            </IconButton>
+            <DeleteDatasetButton dataset={dataset} location="" />
           </PermissionsGate>
         </Box>
       );
     },
-    [askConfirmationToDeleteDialog, confirmAndRefreshDataset, refreshDatasetById]
+    [confirmAndRefreshDataset, refreshDatasetById]
   );
 
   return (
