+++ src/views/DatasetManager/components/DeleteDatasetButton/DeleteDatasetButtonHooks.js
@@ -1,0 +1,11 @@
+import { useDeleteDataset } from '../../../../state/hooks/DatasetHooks';
+import { useWorkspaceData } from '../../../../state/hooks/WorkspaceHooks';
+
+export const useDeleteDatasetButton = () => {
+  const isDatasetCopyEnabledInWorkspace = useWorkspaceData()?.datasetCopy ?? false;
+  const deleteDataset = useDeleteDataset();
+  return {
+    deleteDataset,
+    isDatasetCopyEnabledInWorkspace,
+  };
+};
